move Authenticator::Base to authenticator/base
